- blobby runs under linux_amd64 again - some more cmake fixes

+++ trunk/src/FileRead.cpp
@@ -88,7 +88,7 @@ std::string FileRead::readString()
 {
 	char buffer[32]; 		// thats our read buffer
 	std::string read = "";	// thats what we read so far
-	uint32_t len = length();
+	size_t len = length();
 	
 	while(true)	// check that we can read as much as want
 	{
